PLAT-13089 Refactored ampliude function
